+++ src/constant.ts
@@ -17,3 +17,11 @@ export const CONFIG = {
   CUSTOM_TAGS: "customTags",
   FILES_TO_ENABLE: "filesToEnable",
 };
+
+export const COLORS = {
+  GENERAL_TAG: "#009BE6",
+  CHARACTER_TAG: "#35C64A",
+  COPYRIGHT_TAG: "#C797FF",
+  ARTIST_TAG: "#FF8A8B",
+  META_TAG: "#EAD084",
+};